Update asp-route and form elements in Staff update page
The commit includes a change to use 'name' instead of 'id' for the asp-route and also on the UpdateStaff page form. This enhances UI by hiding the staff 'Id' field, while still keeping it in the form for database operations. It also improves page routing to use more user-friendly identifiers.