Add a new data class and update website UI and functionality
A new data class, Paths.cs, was created to centralize data paths. The website homepage design was significantly updated, and new functionality has been added: the addition of total booking price to the booking model and the display of this information on the AllBookings page as well as the UpdateBooking page. Also, the Json Booking repository and service repository are now using Paths from the new Paths class.