Update UI elements and adjust nullability in models
Updated the styling of the navbar and added a reference to site.css for custom styles. The website logo was also updated. Nullability of properties was adjusted in multiple models to account for potential null scenarios. The Layout for ManageServices page has been specified. Unused imports have been removed from various Interfaces.